requirements.txt
python-telegram-bot
googletrans
aiohttp